The addition of few more 32bit integer into struct snd_mixer causing an accidental MIXER_SIZE overflow. Use 8bit integer instead of 32bit to reduce the kludge size.
Noticed by:	tegge


git-svn-id: svn+ssh://svn.freebsd.org/base/head@162828 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/dev/sound/pcm/mixer.c
@@ -47,9 +47,9 @@ struct snd_mixer {
 	u_int32_t recdevs;
 	u_int32_t recsrc;
 	u_int16_t level[32];
-	u_int32_t parent[32];
+	u_int8_t parent[32];
 	u_int32_t child[32];
-	u_int32_t realdev[32];
+	u_int8_t realdev[32];
 	char name[MIXER_NAMELEN];
 	struct mtx *lock;
 	oss_mixer_enuminfo enuminfo;
